CreateAtom hierarchy problem and Create in atom problem have been fixed.
(When 'in' relation is exist (so in the xml, type of sig is not null and
parent of sig is null), parent of this sig is parent of this type. We
need to change this in all references like Utility.

And there is minor bug when changing atom type. In the xml file, index
of atom in the sig element is wrong )